requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-dotenv==1.0.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
django-allauth==0.59.0
django-rest-auth==0.9.5
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-phonenumber-field==7.2.0
phonenumbers==8.13.27
django-environ==0.11.2
django-storages==1.14.2
razorpay==1.4.1
